fix: make center double to fix panning in higher zoom levels

+++ src/state.h
@@ -4,7 +4,7 @@
 #include "configuration.h"
 
 struct AppState {
-  std::complex<float> zoom_center{conf::kFractalCenter};
+  std::complex<double> zoom_center{conf::kFractalCenter};
   float zoom_scale{conf::kInitialZoom};
   bool center_changed{false};
 };
